keep old legacy package file
